requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
sqlalchemy==2.0.40
argon2-cffi==23.1.0
python-multipart==0.0.20
pytest==8.3.5
httpx==0.28.1